services/streamlit/requirements.txt
# Streamlit Analytics requirements
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
streamlit>=1.30.0
pandas>=2.1.0
plotly>=5.18.0
altair>=5.2.0

